Replace local point with standard image.Point
A 'point' struct was added in 16d439ef to avoid using strings to
represent the points that were visible in the View.Visible map.
This was a good change.

This 'point' is a useful internal representation, but has no exported
fields or methods, so it is of no use to any caller outside this
package. This would not be a problem if external callers never had
access to it, but they do, through the View.Visible map, which is
exported.

This change fixes this by switching from our local 'point' struct to
'image.Point' from the 'image' package in the standard library. This
struct is internally the same, so this change will have no impact in
terms of memory, but it will mean that anyone accessing View.Visible
from outside (eg. for iterating over visible tiles) will now be able
to do things with the stored points.

While this does add a new dependency on 'image', this is in the
standard library, so this should not be a big concern.